pyproject.toml
[project]
name = "ai-vengers"
version = "0.1.0"
description = "AI-powered video production pipeline using Gemini API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.2.0",
    "pillow>=12.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openai"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utils", "examples"]
